# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KONTi Design | Build Studio - Project Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /integrations/drive/folders:
    get:
      operationId: listDriveFolders
      tags: [integrations]
      summary: List Drive folders directly under a parent (defaults to "root")
      parameters:
        - in: query
          name: parentId
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Folders list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DriveFoldersResponse" }
...
  /admin/secrets/{name}:
    post:
      operationId: updateManagedSecret
      tags: [admin-secrets]
      summary: Set or clear an override for a managed secret
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                value: { type: string, description: New secret value (write-only) }
                clear: { type: boolean, description: When true, removes any override and reverts to env }
      responses:
        "200":
          description: Updated row (no raw value)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { $ref: "#/components/schemas/ManagedSecretStatus" }
                required: [status]
  /admin/secrets/{name}/test:
    post:
      operationId: testManagedSecret
      tags: [admin-secrets]
      summary: Probe the live API with the configured key
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Test result (ok flag + safe message)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SecretTestResult" }
  /admin/integrations/restart/{name}:
    post:
      operationId: restartIntegration
      tags: [admin-secrets]
      summary: Force-refresh OAuth tokens and run a health probe (drive | asana)
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Restart result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SecretTestResult" }
...
components:
  schemas:
...
    DriveFolder:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        parents:
          type: array
          items: { type: string }
      required: [id, name]
...
    DriveFoldersResponse:
      type: object
      properties:
        folders:
          type: array
          items: { $ref: "#/components/schemas/DriveFolder" }
        parentId:
          oneOf:
            - { type: string }
            - { type: "null" }
      required: [folders, parentId]
...
    ManagedSecretCategory:
      type: string
      enum: [ai, pdf, presentation, oauth]
    ManagedSecretMeta:
      type: object
      properties:
        name: { type: string }
        label: { type: string }
        labelEs: { type: string }
        description: { type: string }
        descriptionEs: { type: string }
        category: { $ref: "#/components/schemas/ManagedSecretCategory" }
        testable: { type: boolean }
        formatHint: { type: string, nullable: true }
      required: [name, label, labelEs, description, descriptionEs, category, testable]
    ManagedSecretSource:
      type: string
      enum: [override, env, missing]
    ManagedSecretStatus:
      type: object
      properties:
        meta: { $ref: "#/components/schemas/ManagedSecretMeta" }
        source: { $ref: "#/components/schemas/ManagedSecretSource" }
        preview: { type: string, description: "Last-4 masked preview, e.g. …1234" }
        overrideUpdatedAt: { type: string, nullable: true }
        overrideUpdatedBy: { type: string, nullable: true }
      required: [meta, source, preview, overrideUpdatedAt, overrideUpdatedBy]
    SecretTestResult:
      type: object
      properties:
        ok: { type: boolean }
        message: { type: string }
        messageEs: { type: string, nullable: true }
      required: [ok, message]
...